Ported weighed average sql to postgres
